Replace new code base
