+++ .gitignore
@@ -1,0 +1,3 @@
+releases
+go.sum
+tmp
