--- cloudbuild.yaml
@@ -1,11 +1,0 @@
-steps:
-- name: 'gcr.io/cloud-builders/go'
-  args: ['get', '-u', 'github.com/spf13/cobra/cobra']
-  env: ['PROJECT_ROOT=github.com/koduki/hwrap']
-- name: 'gcr.io/cloud-builders/go'
-  args: ['build', '.']
-  env: ['PROJECT_ROOT=github.com/koduki/hwrap', 'CGO_ENABLED=0']
-artifacts:
-  objects:
-    location: 'gs://shared-artifact/'
-    paths: ['hwrap']
