+++ go.mod
@@ -1,0 +1,13 @@
+module hwrap
+
+go 1.13
+
+require (
+	github.com/cosmtrek/air v1.27.3 // indirect
+	github.com/creack/pty v1.1.13 // indirect
+	github.com/fatih/color v1.12.0 // indirect
+	github.com/labstack/echo/v4 v4.3.0
+	github.com/mattn/go-isatty v0.0.13 // indirect
+	github.com/pelletier/go-toml v1.9.1 // indirect
+	golang.org/x/sys v0.0.0-20210531080801-fdfd190a6549 // indirect
+)
